Viewmaster - experimental updateable view creator for Baiji
See the README for more information.  While the work here is
strategically important to Baiji -- exposing a denormalized but
updateable schema -- it is presently orthogonal and not used
by the Baiji web application.

Accordingly, a lot of new libraries are coming along with what
is essentially a one-class generator script.  All this should
merge together soon.

See #42

Change-Id: Iaa3fc09df6dcf8ba20d3171eb93102a824d77d56
Reviewed-on: http://review.source.gogoego.com/330
Reviewed-by: Rob Heittman <rob.heittman@solertium.com>
Tested-by: Rob Heittman <rob.heittman@solertium.com>