References is searchable again.
This commit should make the Reference manager work in the 
client so far as loading and searching is concerned.  Still is looking 
at the wrong database, add/remove functionality has not yet been 
addressed.  This may require some core-level database changes at 
worst, or some query and API changes on the server, at best.

Also, this commit moves ReferenceViewPanel to legacy code, 
LegacyReferenceViewPanel, and is replaced with 
ReferenceViewTabPanel.  I had to slightly rename it due to Git/EGit 
limitations (?).  This makes a cleaner, more manageable, separation 
of concerns between searching and browsing references, and also 
removes the use of many deprecated classes.

Again, while I believe this should satisfy the needs for the ticket, 
there is further work to be done that will be enumerated in 
separate tickets in the near future.

See #11

Change-Id: I2aaf0ae127c300e1a4ab4f5899d7d58e3a6a1048
Reviewed-on: http://review.source.gogoego.com/505
Reviewed-by: Rob Heittman <rob.heittman@solertium.com>
Tested-by: Rob Heittman <rob.heittman@solertium.com>