Workflow compiles.
Added the necessary .hbm.xml files and model objects to make this work. 
Also, replaced the DBSession layer with a Hibernate one.  I make no 
guarantees that the Hibernate persistence layer works in the way that I 
implemented it, and there are a couple methods that have TODOs, but 
I left the DBSession layer in tact so the code is there to go back to for 
testing and comparison.

Also, added a couple utility methods to SISPersistentManager for 
simple object access.

See #71

Change-Id: If5fae3c47cab602cf86ed30b564cf7024dfc7eaf
Reviewed-on: http://review.source.gogoego.com/516
Reviewed-by: Rob Heittman <rob.heittman@solertium.com>
Tested-by: Rob Heittman <rob.heittman@solertium.com>