Refactored structures to take Field model objects
This implementation makes a number of significant changes:

* DominantStructure is now an interface
* DisplayStructure is the interface for Structure -- if the 
method can't be found in DisplayStructure, it's probably 
not being called anywhere anymore and can be removed.
* setData, hasChanged and save have been removed from 
Structure, and only exist in DisplayStructure interface.  
This means that underlying subclasses must implement 
these methods.
* SISPrimitveField is a middle class between Structure and 
a core implementation of a simple Structure, like SISText. 
This houses save, setData, and hasChanged functions, and 
preserves the setData(Map<String, PrimitiveField>) function 
for its subclasses.
* The Ext model object has been removed, thus breaking 
the already broken Classification Schemes.  We can find a 
better way to handle this.
* Anything that did not already have a hasChanged method 
now has one and returns true, such that it always saves.  We 
need to loop back and actually implement these.

With these changes, we haven't lost any functionality, as 
SISPrimitiveFields still save and load as they did before.  We 
have a bit more work to do for Classification Schemes now, 
but the path should be clearer and we are now getting the 
data we need to create the widgets.

See #91

Change-Id: I7487cf482c42a9ede82f6966ca226f9142b4ae63
Reviewed-on: http://review.source.gogoego.com/566
Reviewed-by: Liz Crump Schwartz <liz.schwartz@solertium.com>
Tested-by: Liz Crump Schwartz <liz.schwartz@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>