A number of fixes for loading and saving fields and classification schemes.
Ah, where do I begin?  There are a number of changes that are in this 
commit, but all are related:

* New ClassificationSchemeViewer et al replace the existing widget set 
and remove this code from the ClassificationScheme display for clarity. 
These models do load and save correctly.  The supplied viewer is designed 
to be one-size-fits-many, but there may be other schemes that need 
custom viewers.  The general model supplied, though, should suffice for 
most, if not all, classification schemes.
* Introduced DisplayStructure typing, and differencing between structures 
that are primitive versus ones that aren't.  This is important for saving.
* Modified Field fromXML and toXML functions.  Old assessments from 
disk that do not follow this new structure should still load.
* Added FieldSchemaGenerator for extracting and validating fields and 
field data.
* New shared Debug class can be used for client and server debugging, 
specifically for turning OFF debugging when application is deployed live.
* Existing structures now load and save appropriately, save for SISSelect and 
SISMultiSelect.  These structures need to be revamped to send back the 
proper lookup identifier to the server on save.

See #91

Change-Id: I9bc73ccda85cbec213166c0cc646a72f1aa00590
Reviewed-on: http://review.source.gogoego.com/598
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>