Removal of dead code and diverted System.out to debugging tools.
System.out.println for client and shared becomes Debug.println.  
Right now, there is simple support for a string or object to print, 
as well as String.format-style printing on the client (which is NOT 
natively supported in GWT).  If you're running the client in hosted 
mode, you'll see all printing on the console as normal.  Once built 
and deployed, though, printing stops.  You can also turn printing 
off manually, if you wish, by supplying a different debugger.

On the server, you'll still want to use GetOut, and we can rig things 
in the future to push Debug printing directly to/through GetOut. 

The result -- a quite noticeable boost in speed when debugging 
stops; every little bit helps!

Change-Id: I748f747c5e59e4bb87abb446e8ece8be2dc9e53e
Reviewed-on: http://review.source.gogoego.com/601
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>