Integrity updates to SIS 2.0
This includes implementing the ServerApplication, 
pointing lookups to the sis_lookup database, and 
changing the validation to not use AssessmentTypes.  
This still will not work or validate, but it should be 
able to property popuplate the client with data.

The struct loader generates based on the lookup 
database, as well as the lookup table loader.  This 
should provide users with a similar experience in 
regard to creating Integrity queries, but it does not 
solve the problem yet of actually doing integrity 
validation on the server based on these queries.

See #86

Change-Id: Id2c3391bb238a61b4e66fc9ecd7921af9935c16e
Reviewed-on: http://review.source.gogoego.com/861
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>