Updated hasChanged function to work properly.
For empty assessments, SISBoolean does not understand
the concept of null, so if an assessment does not have
a value, it will be null, but the reported value from
SISBoolean is "false".  Since null and "false" are not
the same, it detects that a change has been made and
auto-saves occur.

This change overrides the hasChanged method to account
for this problem.  If there is no value in the database,
for comparison, it presumes the default value, which is
typically "false", and compares it to the checkbox,
which will either be true or false.  In the latter case,
it will properly detect that changes have not been made.
For assessments with data that contains "false", it
should still work as before.

See #99

Change-Id: Ia804496b06363d46536da2779fdb9f9a309462a5
Reviewed-on: http://review.source.gogoego.com/963
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>