Introducing BaseServiceRestlet!
BaseServiceRestlet should be used pretty much anywhere
we are currently using ServiceRestlet.  This abstract
class provides the following free lunch:

* Request method delegation - simply override functions
to handle the common HTTP methods GET, PUT, POST, and
DELETE.  These methods all throw ResourceExceptions,
which are handled by the caller.  When errors are thrown,
instead of swallowing them, simply wrap them in a
resource exception and throw them.  Simple!  It also
catches your uncaught errors and logs them.
* Entity document parsing - getting a Document from an
XML entity is a common function that needs to be handled
properly but is not always done.  Call a simple function
to handle this work for you.
* Quick representation builder - utility functions provide
you with a means of creating a response document with a
simple message to return.  NativeDocument has built-in
utilities that parse these documents, so you can quickly
send server-generated error or confirmation messages to
the client.

This extends directly from ServiceRestlet, so you still
get all the properties of ServiceRestlet.  In the future,
I would like to branch this away and remove the VFS
component, and the unused AuthorizingRestlet component.

Included in this commit is the conversion of FieldRestlet
to use the BaseServiceRestlet, as an example of how it
can and should be used.

Change-Id: Id808bb39b36d0f6abf889e808f6ca72d205f1591
Reviewed-on: http://review.source.gogoego.com/964
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>