Fixes for the Workflow process, BrowseUsersWindow
Workflow depended on BrowseUserWindow, which was
failing because the underlying Resource was not
actually querying the database properly, as it
referenced old tables.  This change fixes the
users plugin to make the BrowseUsersWindow work
again, which is referenced in multiple places in
SIS, and also fixes the error caused by it in
the workflow process.

Also, fixed some errors in the HBM files for the
Workflow tables, and short-term, made workflow
available offline.  This is untestable otherwise.

See #71

Change-Id: I6da9bfd36fc763b861a75d5042e3552c703ceec0
Reviewed-on: http://review.source.gogoego.com/972
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>