Classification scheme save/hasChanged fix.
Calling save() was wiping the hasChanged marker, 
which is a bad thing.  Changed to pass a boolean 
to specify whether this should happen.  Save now 
performs two tasks:

* Save the models in memory
* if specified, mark the models as not changed.

Since we use save() to also get the models for 
display, needed a way to specify that the hasChanged 
flag NOT be modified.

Change-Id: I9bf477a5ddae7085fb378cdfc1b83d0dd0bbae18
Reviewed-on: http://review.source.gogoego.com/990
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>