Added saveIfNecessary to AssessmentClientSaveUtils
This function checks to ensure that the current
assessment doesn't need to be saved.  This is
now called by default before changing a working
set, and also a number of places in the navigation.
This could be pushed down further, if warranted.

This is the same functionality from the Quick
Taxon browsing, so it's been proven to work.

See #100

Change-Id: I983f77f1e48cc295f8a0d3f933135f5b4a019c3d
Reviewed-on: http://review.source.gogoego.com/1001
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>