Fixed javascript error in TaxonomyBrowserPanel
Seems the problem was that functions were not 
being called in a synchronous manner, which 
explains why the problem did not consistently 
manifest.  Fix simply called for a bit of code 
cleanup, forcing operations to be synchronous 
and using proper callback patterns.

In addition, did some code cleanup to make the 
panel follow the listener model (free lunch from 
Ext) and cleaned up classes that were in-line 
overriding this where it was feasible/useful to 
do so.

Finally, killed off unnecessary calls to layout() 
which should improve render speed.

See #97
See #127

Change-Id: Id62e518911376835c024708a71ef8a381713ba70
Reviewed-on: http://review.source.gogoego.com/1015
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>