SIS 2.0 implementation of locking.
While there is a model for Locking, I decided not
to use it for expediency, opting to simply
re-implement the PersistentLockRepository.  The
HibernateLockRepository is implemented, though.  I
am not sure we need a model here either, but one is
in place regardless.  It is trivial to switch from
one locking implementation to another, so not a big
concern.

Also, I simplified the data model by coalescing the
persistentlock and persistentlockgroup tables into
one lock table, adding a group column.  By doing this,
I was able to change the client to only make one
call to the server to render the table instsead of
two.

See #128

Change-Id: I7ace7563e7d9a37b18d39bb6e23ceb91e396effd
Reviewed-on: http://review.source.gogoego.com/1020
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>