Multi-select hasChanged fix.
For some reason, Ext is reverting our coded
list item IDs to default auto-generated IDs,
so when a page is re-generated, the field IDs
are now different; thus, leaving a page causes
it to invoke save behavior.

Documentation says this should only happen if
the ID is unset, but we are setting it, so
instead, I am working around the issue by
using another means of storing our system
identifiers.

See #99

Change-Id: Ic475a77ce7b2be0c3826f54446b1557d6639af7c
Reviewed-on: http://review.source.gogoego.com/1035
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>