Taxon Home Page re-vamping and cleanup.
Taxon home page was doing a lot of extra re-drawing,
causing increased load times unnecessarily, due to
other system listeners.  Eliminated these extra
refreshes.

Also, a lot of layout() calls happening, which also
slows things down.  Because this panel uses a lot of
core GWT widgets that don't support layout(), this is
not helpful.  Removed some of these.

These layout calls were are result of re-using the
same panel when drawing a new taxon.  However, there
is no need for this because the behavior did not
re-use any widgets (probably the intention, in which
case layout() is necessary), but rather, it removed
everything and re-drew it.  Changed this so it redraws
everytime, but does not attempt to layout, lets
natural layout occur on render.

Finally, updated the call on the server to retrive
Taxon info fresh instead of from cache, as the cache
is out-dated since latest Common Name changes.

With more predictable and efficient behavior,
rendering issues should be resolved.

In regard to 142, added clickable style to the
hierarchy icon, so mouseover produces indication
that the image is clickable.

See #142

Change-Id: Id76439d23d918fe982f64773030cd6875247547f
Reviewed-on: http://review.source.gogoego.com/1103
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>