Monkey Navigator 2.0 selection fixes.
This involved updating how the cache responded to changes,
ensuring appropriate setting and nulling were being done.
The updating of the navigator was happening correctly, but
the calls weren't always making it there.  Updated the
necessary caches to fix this.

Added scroll into view functionality for taxon listing,
and fixed issue with selected Working Set and selected
Taxon not being displayed as selected when the navigator
renders.

See #143
See #145

Change-Id: I5623d95db65f6cc1c488950e3834c0d323327571
Reviewed-on: http://review.source.gogoego.com/1104
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>