Better behavior saving note/ref to unsaved fields.
When attempting to write a note or assessment to a
field without an ID, it will automatically add the
field to the assessment and save it, generating an
ID for the field, before popping the interface for
adding notes/references.

This may need to be re-visited for classification
schemes, but should work for all other fields.

See #125

Change-Id: I436a81b2aada7e871358f5254272030128af4282
Reviewed-on: http://review.source.gogoego.com/1106
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>