Remove ghost taxon home page, fix dual loading.
Multiple calls to draw were causing the taxon
home page to render a second time when a
certain path was used to render the page
(using the taxon children panel).

This commit removes the explicit redraw call
when changing children, allowing natural redraw
behavior to be called based on other listeners
in the system.

Additionally, added a removeAll call in tehe
description panel such that it is impossible
for there to be two panels, even IF it were
to call draw twice (which it should not).

See #149

Change-Id: Ia84d17a9f1a4e3758604a5115bb869df0df4960f
Reviewed-on: http://review.source.gogoego.com/1127
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>