Structure Improvements for RegionInformation, Date
RegionInformation had some CSS failures as well as
some poor widget use that was hurting the UI,
probably due to the Ext upgrade.  Fixed those.

Also, for Date, incoming dates are timestamps, but
they are saved as formatted dates.  This causes
hasChanged to be true, thus an unwarranted save.
Overrode the hasChanged method to deal with this
appopriately.

See #134

Change-Id: I36919a8cb4a537c479c4b4271e6d7e6ba38b379a
Reviewed-on: http://review.source.gogoego.com/1191
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>