Fix bugs with working set load/save
Working sets were not correctly saving or loading
data in the edit working set or new working set
panels.  Specifically, regions were not being
handled correctly.

Added code to correctly persist regions on the
server, and fixed bugs with writing and loading
the correct set of regions on the client.

Also, updated the models to parse correctly, which
also meant adding equals/hashcode to Region and
User models.

Fixed a delete bug that was causing the
user authentication window to popup.

See #122
See #179

Change-Id: I38f27b949760f4cf60c74821d878ca2d11429422
Reviewed-on: http://review.source.gogoego.com/1269
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>