Standardized Note Viewer, Synonym Notes
Standardized the way notes are handled across
all notable objects, including Taxa, Displays,
Common Names, and Synonyms.  Essentially,
pushes all the custom code that does interesting
stuff into a NoteAPI that handles calls, and
the viewer is the same across the board.

In doing this, it also fairly trivially fixed
notes for Taxa and Synonyms.

See #150

Change-Id: I68e6ee968cd79536d7ee68341cae159c914ca358
Reviewed-on: http://review.source.gogoego.com/1273
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>