Checkable Search Panel, SearchPanel Bug Fix
Added an explicit implementation of a checkable
search panel, for which results display a check
box in the table to designate selections. Later,
this should be ported to a Grid, but works fine
now.

Also, fixed a possible NPE in search panel if an
event was called to fire without at least
one listener.

Change-Id: I08c2c27d129babd98611d629e6660ed08f632df8
Reviewed-on: http://review.source.gogoego.com/1274
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>