Read IsoLanguages from Database, LanguageCache
Ported IsoLanguages from the VFS to the database.
Added a LanguageCache implementation to the client
to handle access and update of languages, though
the server only handles GET.

TODO here is to cache the resulting listing XML
on the VFS such that it's not re-fetched each time,
but that can happen later.  Probably need a means
of updating this table as well, and clearing the
VFS cache on update.

The win here is for data porting -- if we forget
to copy over the language XML file, common names
will not work.  Now, this file is not needed, so
SIS should function correctly directly from a
data conversion, without any extra file copying.

See #162

Change-Id: Ieebc7a19847675b1ccb0cb526bf61d4ae7be33c3
Reviewed-on: http://review.source.gogoego.com/1282
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>