Removed unnecessary call to center(), caused error
Calling Window.center on some windows could cause
a javascript error. This was actually simply an
assertion error that did no harm, but I don't
want to rely on "responsible implementations" for
the assertion.  For windows that render with
asynchronous patterns, the center() call will
bomb.  Instead, just removed it.

Change-Id: Ief33c472302f08fc851fc11a1d3a07f5a8e0b3f1
Reviewed-on: http://review.source.gogoego.com/1286
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>