Bug Fix: Undo checks for wrong history text
History text is properly amended with the
"as" portion not italicized, but the undo
was not checked for "as" in a comparison
that unset this value.  Changed the statement
so that it does.

While the fix is related, this does verify
that the text is as it should be.

Test #65

Change-Id: If13bdde632f0239de9d70e5e55fec7c2d5660674
Reviewed-on: http://review.source.gogoego.com/1293
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>