Bug Fix: Class.Scheme setData check for primitive.
Classification schemes's setData function was not
checking to see if the default structure was a
primitive field or not.  Added this check.

Additionally, a couple classification scheme XML
definitions were incorrect, specifying bogus
identifiers on some structures, which would
break the model.  Removed these unnecessary
identifiers.

By not having these checks, the wrong fields
will be passed to structures that are expecting
different field types, resulting in ClassCast
errors.

See #187

Change-Id: Ifd41d79d20d07707919b3343a576096ff6268820
Reviewed-on: http://review.source.gogoego.com/1307
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>