Bug Fix: Users not saving/loading nickname via toFullXML
UserRestlet checked explicitly for certain fields that
could be updated, but did not fail if an invalid field
was being saved, causing a 200 and the client to believe
that their data had been saved.  Changed this so that if
a field that does not exist in the database is passed,
a failure response is returned.

Added a check for the new nickname field here.  We should
probably do something smarter in the future for this
function to have longevity.

Finally, added nickname to the toFullXML function.  Was
previously only being written in the toXML function.

See #186

Change-Id: I438296f591b8719f16cd8b682076ad03fea2ddac
Reviewed-on: http://review.source.gogoego.com/1311
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>