Bug Fix: ClassCastException in Class.Scheme Model
If the structure was a primitive structure, a
class cast exception was being thrown, since the
field is always passed to setData.  Fixed this
by detecting this and passing the appropriate
field.  This fix is in addition to the previous
fix, since while the changes applied worked, the
setData function was being called again down
the line, causing the error to continue to occur.

See #187

Change-Id: I4cf7d74821f9afd4ff149420948b911b44e2478a
Reviewed-on: http://review.source.gogoego.com/1312
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>