Fix ReferenceEditor for new/existing refs, server.
Changed the XML writing in References to support
optional fields, and only write to XML those fields
that exist.  Serializer ensures that null values
are never written as "null", which causes deserialization
errors.  Also, added a few static helper methods.

Fixed a XML writing error in the SubmissionResource that
caused the entity to report bad XML, and the client to
be unable to parse the entity.

Added double-click handling on the Reference search page
to open the selected reference in the ReferenceEditor.

See #188

Change-Id: I596ea680e96b806a69deb6925973f36454e8fb64
Reviewed-on: http://review.source.gogoego.com/1319
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>