Bug Fix: Navigation only prompts for save once.
To make functionality consistent, the data caches
check to ensure an assessment is saved before
changing status.  For the navigation, we want to
do this check prior to other work, and disable
the additional call.

Added functions to cache APIs where necessary to
allow this overriding, and changed the navigation
to call the appropriate functions.

See #154

Change-Id: Ic0eb4551bac193137e4a8edf814b51c0e3dab68b
Reviewed-on: http://review.source.gogoego.com/1320
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>