Bug Fix: Displays not displaying/updating toolbar correctly.
When moving from one assessment to another, the icon panel
toolbar would not update it's display to show the correct
status for notes/references.  Added a call to setData at
the Display level to take care of this.

Also, fixed a small loophole with the reference icon that
could allow it to be the wrong color.

Change-Id: Ib581bc9eaff4282f0e8e44893da91be1d06d9c9b
Reviewed-on: http://review.source.gogoego.com/1321
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>