Reference load and display changes for viewing panel.
Bibliography tab displays the citation, but for some
reference types, the citation was not being generated.
This caused the bibliography panel to appear empty even
if it wasn't.  Added a function to ensure that something
always appears when that panel is displayed.

Also, some reference types were not being correctly parsed.
Changed the parser to use a better matching pattern to find
all supported types.

Finally, for types that aren't supported, default to "Other"
as a type.  Added a FIXME specifically for the type
"Personal Communication," which is not supported as of right
now, but looked to be in the past.

See #189

Change-Id: I9d3b6c21c3409fd99eee2b7c4d61f9c5242ef455
Reviewed-on: http://review.source.gogoego.com/1322
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>