Recent Assessments Panel speed improvements
Pushed the recent assessment panel's data processing
from the client to the server for expediency.  The
information from the server is now simply a literal
translation of what's seen in the grid.  Assessment
information in full isn't transferred to the client
until an assessment is clicked.

Also, cleaned up the AssessmentCache code a bit and
changed the RecentAssessmentRestlet to extend
BaseServiceRestlet.

Since recent assessments gets loaded for every user
upon login, and the list can get long, this should
provide an immediate speed boost for initial login.

Change-Id: Ib46307307d0a237502a0d21a1a8aa7165fe31a63
Reviewed-on: http://review.source.gogoego.com/1327
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>