Layout and styling for BugPanel, DEM
Updated the layout of the BugPanel to be consistent
with the recent assessment panel, added additional
text and styling options to the DEM View, including:

* h1
* h2
* h3
* information
* warning

The "text" option has been left untouched.  We can
update this style later.  h3 may be appropriate for
most places where text is being used.

Change-Id: I11d90291fd25f238a67184024b653d90ee49c9e0
Reviewed-on: http://review.source.gogoego.com/1328
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>