User Loading Fix, Changed User HBM definition.
Users were not loading all the necessary fields
to render to UserViewPanel, specificially the
nickname field and the permissions fields.

Changed the custom SQL query to use Hibernate
layer, and, more importantly, changed the HBM
definitions for User to define permissions as
non-lazy.  This could affect other areas of
the system, but was needed here.  Should not
adversely affect other parts of the system, but
it should be noted nonetheless.

Also, changed UserRestlet from a ServerResource
to a BaseServiceRestlet.  While a technical
regression, in practice, no one ever actually
uses ServerResource.  Rather have consistency
over freshness.

See #186

Change-Id: Ie6c0e03d59138b5b4035125b6fcdf9fc6087e1bb
Reviewed-on: http://review.source.gogoego.com/1330
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>