Client/Server Assessment Schema Plumbing
The assessment schema defines the views and
fields that a particular assessment uses to
capture data.  The default schema is the
Red List Assessment schema, but using GoGoEgo
brokering fundamentals, any assessment
schema can be dropped in and used (almost)
on the fly.

If no schema is provided, code currently
assumes the Red List Assessment schema,
though later we should make this an application
setting.

Added server plumbing to support all this, and
a means via the New Assessment Panel to create
a new assessment and assign it a schema.

Change-Id: Ibb2ff6c0bbe4632a9c169b12f0ff390bb3159b25
Reviewed-on: http://review.source.gogoego.com/1352
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>