Field Widget URL Fix, Schema Table Prefixes
Fixed a bad URL in the FieldWidgetCache.

Added utility method to PrimitiveFieldType.

Added default table prefix for the AssessmentSchema,
such that all fields in the schema will begin with
the prefix.  RedListAssessment has a blank prefix.
This is just for quick-glance purposes, should have
no significant effect on coding, and no assumptions
should be made based on the prefix.

See #48

Change-Id: I66f2a75b53e562ea14bd9c1913735917d227cb01
Reviewed-on: http://review.source.gogoego.com/1380
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>