Resurrected Tree-based coding option selection.
Added the full-add button back to Classification
Schemes, porting the tree to the new Ext TreeStore
and BaseTreeModel widgets, removing deprecation.

Changed the BasicClassificationSchemeViewer API a
bit to allow for the selected row to be passed to
the widget generator.  Threats will override this
to generate the correct editing widget.

See #205

Change-Id: I2e88b12a1626aa57d2833bcc649e8168f3187643
Reviewed-on: http://review.source.gogoego.com/1405
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>