Server-configurable default assessment schema
Added a means to configure the default assessment
for a particular server instance.  The default
schema is managed by the SIS application's settings
page and can be customized for a given deployment.

All calls that used Assessment.DEFAULT_SCHEMA have
now changed to be SIS.get().getDefaultSchema() on
the server and SchemaCache.impl.getDefaultSchema()
on the client.  The client now pulls the default
schema setting from the server.

In addition, fixed some bugs related to switching
from schema to schema in the assessment data
browser, and exposed the red list assessment
schema in the manifest.

Change-Id: I8b36141a2bd07d2207af01c5680f8f6224905c8c
Reviewed-on: http://review.source.gogoego.com/1409
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>