Use/Trade View, Detailed Wild Harvest Record
Implemented basic fields for Wild-Harvest record
into the correct model, updated the field model
for this field as well, but we will only need this
for testing purposes now.

Field consist of a one-to-many field with a nested
one-to-many field for end-use records.

Also, added view with accurate field names.

See #221
See #222
See #223

Change-Id: Iad2070d7eb97d2df49852495ffbf158a42d75d1f
Reviewed-on: http://review.source.gogoego.com/1458
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>