Amend permissions to allow assessment schema selection
Added a drop-down to permissions such that the schema
can be interrogated.  One can choose a particular
schema from the list, or all schemas.

Particularly, this will allow for permissions to be
created such that users can only edit a particular
type of assessment, such as a use/trade assessment.

See #213

Change-Id: I63e2223705e8dd98d7adebbd568f5b9ddd9fcf6e
Reviewed-on: http://review.source.gogoego.com/1546
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>