Remove close icon from ClassificationSchemeViewer
You must noweither click "Done" or "Cancel".  Cancel
actually does change reversion, while Done perists
the changes.  Closing the window with the 'X' does
nothing.

While this may not be the reported issue (not specified
in the bug report), (a) I suspect it is, and (b) this
is still an issue that could arise in the future, so
taking care of it now.

See #235

Change-Id: I5088da411ea0933aea228b3a0e86367bb69c7e64
Reviewed-on: http://review.source.gogoego.com/1550
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>