Build Number Support on Server, Client; Themes.
Added support for tracking a build number with
each SIS client build.  This allows us to view
past builds of SIS as well as help with caching
and communication problems by ensuring that
users and testers are on the same build of SIS
when reporting issues.

Also, fixed the theme support and added missing
themes to the project and the build path.

Change-Id: I33e298058bedf57501885e3a2f84cb28359e0a08
Reviewed-on: http://review.source.gogoego.com/1622
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>