Bug Fix: TrashRestlet patches to make it more accepting.
TrashRestlet was blowing up if an Edit could not be found.
Some data may be deficient for some reason, so I've added
code that can handle this deficient data, instead of
failing with a NPE for the entire call.

These unknown values can be handled by the client in a
different manner later, if desired, though I think the
data omission in question doesn't really matter.

Change-Id: I6fc4f8d977b7bbff2ba57505dd2ccee62586e6ce
Reviewed-on: http://review.source.gogoego.com/1698
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>