Bug Fix: Only select template assessments within schema.
Fixed a bug that allowed users to create assessments
in one schema (i.e. Use/Trade) based off of template
assessments that were from another schema (i.e RedList).
This is not useful.

Also, in a related note, updated the Assessment object
such that accessing the RegionField's properties now
goes through the RegionField proxy, providing a
standard and predictable way to get at the data that
may or may not be there.

See #236

Change-Id: I50d47db94f9470c102d9b9af340952fe1520c6c9
Reviewed-on: http://review.source.gogoego.com/1716
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>