Overdue client-side code cleanup
First round of code cleanup.  Removed dead code and unused
classes in anticipation of the UI changes (don't want to
deal with sifting through code we don't use), and also
because it's a good thing to do in general.

Eclipse should show no warnings for org.iucn.sis.client
and org.iucn.sis.client.shared.

Warning have been removed by either fixing the cause or,
in many cases, adding SuppressWarnings flags.  Typically,
these are used in the following cases:

"deprecation" - when Trees, Tables, or other complex
structures are used, such that swapping them out right
now is a Task.  Also, DataList is deprecated, but its
replacement, ListView, has known bugs, so not changing
over until that's cleared up.

"unchecked" - typically for Field and PrimitiveField.

"unused" - for code blocks that are known to be unused
for a particular, specified reason, OR it has been
removed in the short-term and will be added back later.

Change-Id: I0f21f49f3a701560cedd143379c7a7bcf9668a26
Reviewed-on: http://review.source.gogoego.com/1734
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>