JUnit dependency on testing project, Batch Change
Added a JUnit 4 dependency on the testing project,
which should be fine because it never actually gets
deployed to any server.

Additionally, set up a test source folder with a
couple packages for housing test cases for core
SIS functionality and another for SIS extensions.
Also, addeed a stub class simply to avoid Git
not recognizing the package.

Finally, added an inaugural green test case to
test the Batch Change operation.

Will look to move some of our existing test
cases to this model in the future.

See #90

Change-Id: I4033e7ab22a1b7a5753718200f1d8cd5868614f7
Reviewed-on: http://review.source.gogoego.com/1759
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>