Moved definitions from file system to database.
Definitions are now stored in the 'definition' table in
the Postgresql database.  This requires the addition of
a simple table, which you can generate with the SQL below:

   CREATE TABLE definition
   {
     id integer NOT NULL,
     "name" character varying(255) NOT NULL,
     "value" text,
      CONSTRAINT definition_pkey PRIMARY KEY (id)
   }

Also, changed the definition UI and definition cache to
use the new Definition model and load/save correctly.

Finally, added a DefinitionConverter to the conversions
project that will take the defintions currently stored
on the file system in SIS 1, load them up, and save them
as definitions in the database in SIS 2.

Change-Id: I072744f52e5c74d3c0ceb6049d6463ad3f298176
Reviewed-on: http://review.source.gogoego.com/1771
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>