Recently Accessed Object fixes, User support
Added some fixes in that should have gone into
the previous commit... :-\ -1 git branch

Also, added server and client support for having
recent users.  The recent user list is downloaded
upon login (which is not useful but gets a result).

Finally, added real support in the BrowseUsersWindow
for loading recent users after logout, as the list
is now persisted on the server.  In a somewhat
related note, the ProfileSearchResource now also
lists the nickname in the result set; although this
is not currently displayed, went ahead and added it
in.

See #191
See #232

Change-Id: I46e08a69832099cd0f5bb916f0b168def3f57bf0
Reviewed-on: http://review.source.gogoego.com/1796
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>