Support for RedListFacilitators field.
Need to get this added to the lookups
database via the provided script, but
otherwise all should be covered by this
commit.

Also, fixed the SISOptionsList variables
to accept the right primitive field names.
Was looking for "textValue" and "users",
but the SQL declarations said that the
names were "text" and "value", respectively.

Test #190

Change-Id: I1d44c36614215f9a2c1745b5e4d1f2b4bc90b387
Reviewed-on: http://review.source.gogoego.com/1835
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>